create wire user repository
