+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/user/UserModel.kt
@@ -8,17 +8,6 @@ abstract class User {
     abstract val id: UserId
 }
 
-data class WireUser(
-    override val id: UserId,
-    val name: String?,
-    val handle: String?,
-    val email: String? = null,
-    val phone: String? = null,
-    val accentId: Int,
-    val team: String?,
-    val previewPicture: UserAssetId?,
-    val completePicture: UserAssetId?
-) : User()
 
 data class SelfUser(
     override val id: UserId,
