+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/wireuser/search/WireUserSearchResult.kt
@@ -1,6 +1,6 @@
-package com.wire.kalium.logic.feature.user.search
+package com.wire.kalium.logic.feature.wireuser.search
 
-import com.wire.kalium.logic.data.user.WireUser
+import com.wire.kalium.logic.data.wireuser.model.WireUser
 
 //TODO: this model could be extended later on if we want to include more info into the search result
 data class WireUserSearchResult(val wireUsers: List<WireUser>)
