+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/feature/user/SearchPublicUserUseCaseTest.kt
@@ -5,8 +5,8 @@ import com.wire.kalium.logic.data.publicuser.PublicUserRepository
 import com.wire.kalium.logic.data.publicuser.model.PublicUser
 import com.wire.kalium.logic.data.publicuser.model.PublicUserSearchResult
 import com.wire.kalium.logic.data.user.UserId
-import com.wire.kalium.logic.feature.user.search.SearchPublicUserUseCase
-import com.wire.kalium.logic.feature.user.search.SearchPublicUserUseCaseImpl
+import com.wire.kalium.logic.feature.wireuser.search.SearchPublicUserUseCase
+import com.wire.kalium.logic.feature.wireuser.search.SearchPublicUserUseCaseImpl
 import com.wire.kalium.logic.functional.Either
 import io.mockative.Mock
 import io.mockative.anything
